Moved login.js
